+++ R/Projects.R
@@ -40,6 +40,8 @@ DeleteProject <- function(project) {
 #' @param filter list. Optional. A named list that can be used to specify various filters.
 #'  Currently `projectName` is supported which will filter returned projects for projects with
 #'  names containing the specified string.
+#' @param limit integer. Optional. At most this many results are returned, default: 1000
+#' @param offset integer. Optional. This many results will be skipped, default: 0
 #'
 #' @return An S3 object of class 'projectSummaryList', consisting of the following elements:
 #' \itemize{
@@ -59,8 +61,6 @@ DeleteProject <- function(project) {
 #'     project.
 #'   \item partition. Dataframe with one row for each project and 12 columns specifying
 #'     partitioning details.
-#'   \item recommender. Dataframe with one row for each project and 3 columns characterizing
-#'     recommender projects.
 #'   \item advancedOptions. Dataframe with one row for each project and 4 columns specifying values
 #'     for advanced option parameters.
 #'   \item positiveClass. Character string identifying the positive target class for binary
@@ -80,9 +80,9 @@ DeleteProject <- function(project) {
 #'   ListProjects(filter = list("projectName" = "TimeSeries"))
 #' }
 #' @export
-ListProjects <- function(filter = NULL) {
+ListProjects <- function(filter = NULL, limit = 1000, offset = 0) {
   routeString <- "projects/"
-  params <- NULL
+  params <- list(offset = offset, limit = limit)
   if (!is.null(filter)) {
     if (!is.list(filter)) {
       stop("`filter` must be a list.")
@@ -91,7 +91,7 @@ ListProjects <- function(filter = NULL) {
       if (length(filter$projectName) != 1) {
         stop("`projectName` must be a character vector of length 1.")
       }
-      params <- list("projectName" = filter$projectName)
+      params$projectName <- filter$projectName
     }
   }
   returnValue <- DataRobotGET(routeString, query = params)
@@ -100,7 +100,8 @@ ListProjects <- function(filter = NULL) {
 
 
 projectSummaryList <- function(projectSummaryData) {
-  emptyProjectSummaryList <- structure(list(projectId = character(0), projectName = character(0),
+  if (length(projectSummaryData) == 0) {
+    emptyProjectSummaryList <- structure(list(projectId = character(0), projectName = character(0),
                  fileName = character(0), stage = character(0), autopilotMode = logical(0),
                  created = character(0), target = logical(0), metric = logical(0),
                  partition = data.frame(datetimeCol = logical(0), cvMethod = logical(0),
@@ -109,9 +110,6 @@ projectSummaryList <- function(projectSummaryData) {
                                         userPartitionCol = logical(0),  validationType = logical(0),
                                         trainingLevel = logical(0), partitionKeyCols = logical(0),
                                         holdoutPct = logical(0), validationLevel = logical(0)),
-                 recommender = data.frame(recommenderItemId = logical(0),
-                                          isRecommender = logical(0),
-                                          recommenderUserId = logical(0)),
                  advancedOptions = data.frame(blueprintThreshold = logical(0),
                                               responseCap = logical(0), seed = logical(0),
                                               weights = logical(0)),
@@ -119,15 +117,13 @@ projectSummaryList <- function(projectSummaryData) {
                  holdoutUnlocked = logical(0),
                  targetType = logical(0)),
             class = "projectSummaryList")
-
-  if (length(projectSummaryData) == 0) {
-    emptyProjectSummaryList
+    return(emptyProjectSummaryList)
   } else {
     idIndex <- which(names(projectSummaryData) == "id")
     names(projectSummaryData)[[idIndex]] <- "projectId"
     projectSummaryData <- as.dataRobotProject(projectSummaryData)
     class(projectSummaryData) <- "projectSummaryList"
-    projectSummaryData
+    return(projectSummaryData)
   }
 }
 
@@ -151,7 +147,6 @@ projectSummaryList <- function(projectSummaryData) {
 #'   \item metric. Character string specifying the metric optimized by all project models.
 #'   \item partition. A 7-element list describing the data partitioning for model fitting
 #'     and cross validation.
-#'   \item recommender. A 3-element list with information specific to recommender models.
 #'   \item advancedOptions. A 4-element list with advanced option specifications.
 #'   \item positiveClass. Character string: name of positive class for binary response models.
 #'   \item maxTrainPct. The maximum percentage of the project dataset that can be used without going