+++ R/UploadData.R
@@ -6,9 +6,25 @@
 #'
 #' @seealso SetupProject
 #' @param dataSource character. The file to upload.
+#' @param fileName character. The name of the file after it is uploaded. If not set, defaults
+#'  to the name of the uploaded file.
 #' @return An httr object specifying the form upload content of the file path.
-UploadData <- function(dataSource) {
-  httr::upload_file(DataPathFromDataArg(dataSource))
+UploadData <- function(dataSource, fileName = NULL) {
+  if (!is.null(fileName)) {
+    stopifnot(is.character(fileName))
+  }
+  dataPath <- DataPathFromDataArg(dataSource)
+  content <- httr::upload_file(dataPath)
+  if ("name" %in% names(content) && is.null(content$name)) {
+    # DSX-2862
+    # httr::upload_file invokes curl::form_file.  In curl<5.0.1, this function
+    # returns a length-2 list containing $path and $type. In curl==5.0.1, thanks to
+    # https://github.com/jeroen/curl/pull/290 this function returns an extra $name.
+    # To ensure this works properly with `datarobot:::TryingToBeNull()`, set $name
+    # to the path.
+    content$name <- ifelse(is.null(fileName), yes = basename(dataPath), no = fileName)
+  }
+  invisible(content)
 }
 
 #' Get the data path.